add content type at frontend lib [skip ci]
