+++ frontend_lib/src/helper.js
@@ -288,3 +288,12 @@ export const IMG_LOAD_STATE = {
   LOADING: 'loading',
   ERROR: 'error'
 }
+
+// INFO - CH - 2019-06-11 - This object must stay synchronized with the slugs of /api/v2/system/content_types
+export const CONTENT_TYPE = {
+  HTML_DOCUMENT: 'html-document',
+  FILE: 'file',
+  THREAD: 'thread',
+  FOLDER: 'folder',
+  COMMENT: 'comment'
+}
